% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_families.R
\name{get_families}
\alias{get_families}
\title{Get taxa names from a single web page, or multiple pages, on FNA.}
\usage{
get_families(from = NULL)
}
\arguments{
\item{from}{Source flora, one of 'fna', 'chile', 'china', 'china_moss', 'nepal',
'missouri', 'ecuador', or 'jepson'}
}
\value{
Taxa names in a vector or data.frame
}
\description{
Get taxa names from a single web page, or multiple pages, on FNA.
}
\examples{
\dontrun{
get_families("fna")
get_families("chile")
get_families("china")
get_families("china_moss")
get_families("nepal")
get_families("missouri")
get_families("ecuador")
get_families("jepson")
}
}
