% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genera.R
\name{get_genera}
\alias{get_genera}
\title{Get generic names for a family.}
\usage{
get_genera(from = NULL, family = NULL, fuzzy = FALSE, ...)
}
\arguments{
\item{from}{Source flora, one of 'fna', 'chile', 'china', 'china_moss', 'nepal', 
'missouri', 'ecuador', or 'jepson'}

\item{family}{A plant family name.}

\item{fuzzy}{If TRUE, does fuzzy search using \code{\link{agrep}}. If FALSE, uses
\code{\link{grep}}.}

\item{...}{Further args passed on to \code{\link{agrep}} or \code{\link{grep}}}
}
\value{
Taxa names in a vector or data.frame
}
\description{
Get generic names for a family.
}
\examples{
\dontrun{
out <- get_genera(from='fna', family='Asteraceae')
out$names
head(out$urls)

# Search for trait data
receptacle(out$urls$url[3])
}
}
